Moved functionality from application controller to concerns
